two_todo/django_todo_api/requirements.txt
# framework dependencies
django
django-seed
djangorestframework
djangorestframework-api-key
django-cors-headers


# python dependencies
pyyaml
requests
python-dotenv
psycopg2-binary

djangorestframework-simplejwt
